requirements.txt
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
gunicorn==19.7.1
Pillow==5.1.0
pytz==2018.4
dj-database-url==0.4.2
psycopg2==2.7.4
pyodbc==4.0.23
WeasyPrint==0.42.2
requests==2.18.4
